+++ src/NumArray/FactoryInterface.php
@@ -7,4 +7,6 @@ interface FactoryInterface
     public function createFromData(array $data): NumArray;
 
     public function createZeros(int ...$axis): NumArray;
+
+    public function createZerosLike(NumArray $numArray): NumArray;
 }